ai-service/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
google-generativeai
requests
python-dotenv
numpy
